rag-backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langgraph
chromadb
pypdf2
openai
huggingface_hub
unstructured